refactor(rsc): rename
