+++ packages/rsc/src/core/browser.ts
