+++ packages/rsc/src/extra/ssr.tsx
@@ -5,10 +5,7 @@ import ReactDOM from "react-dom";
 import type { ReactFormState } from "react-dom/client";
 import ReactDomServer from "react-dom/server.edge";
 import ReactClient from "react-server-dom-webpack/client.edge";
-import {
-  createServerConsumerManifest,
-  setRequireModule,
-} from "../core/client-ssr";
+import { createServerConsumerManifest, setRequireModule } from "../core/ssr";
 import { importAssets } from "../ssr";
 import type { RscPayload } from "./server";
 import {
