+++ packages/rsc/tsdown.config.ts
@@ -6,10 +6,9 @@ export default defineConfig({
     "src/browser.ts",
     "src/ssr.ts",
     "src/rsc.ts",
-    "src/core/client-browser.ts",
-    "src/core/client-ssr.ts",
-    "src/core/server.ts",
-    "src/core/shared.ts",
+    "src/core/browser.ts",
+    "src/core/ssr.ts",
+    "src/core/rsc.ts",
     "src/extra/server.tsx",
     "src/extra/ssr.tsx",
     "src/extra/browser.tsx",
